MacPatch 3.3.2 Bug Fix Release - Fixes local database - Fixes tasks rerunning every 30seconds - Fixes SUS Server reset

+++ Source/MacPatch/MPLibrary/MPLibrary/MPRESTfull.m
@@ -71,7 +71,7 @@
     
     req = [[MPHTTPRequest alloc] init];
     wsresult = [req runSyncGET:urlPath];
-	qldebug(@"[getDataFromWS][result]:%d",wsresult.statusCode);
+	qldebug(@"[getDataFromWS][result]:%ld",(long)wsresult.statusCode);
     if (wsresult.statusCode >= 200 && wsresult.statusCode <= 299) {
         logit(lcl_vDebug,@"Get Data from web service (%@) returned true.",urlPath);
         logit(lcl_vDebug,@"Data Result: %@",wsresult.result);
